Add Vs2019 support
Adds Vs2019 support
 * only notable difference between 2017 and 2019 is that the ToolsVersion attribute in the header was moved to its own Property value in the Globals PropertyGroup.

* Fix some warnings in the premake C code.
* Fix VS solution configuration to properly compile against x64.
* Add a gitignore or the ".build" folder which is what i was using for generating the VS solution into.

+++ src/host/premake.c
@@ -265,7 +265,7 @@ int process_option(lua_State* L, const char* arg)
 	char* ptr = strchr(arg, '=');
 	if (ptr)
 	{
-		int len = ptr - arg;
+		int len = (int)(ptr - arg);
 		if (len > 511) len = 511;
 		strncpy(key, arg, len);
 		key[len] = '\0';
